requirements.txt
PySide6
SQLAlchemy
PyMySQL
pandas
numpy
Faker
matplotlib
networkx
scikit-learn
xgboost
ortools
python-dotenv
python-docx
pillow
